Make everything iterators (#19)
This adds an iterator chain from the transactions into the stream of
batches which get loaded into the batch queue. I have added an iterator
adapter for the channel bank which takes in an iterator of frames.
Because iterators are properly lazy the ordering that channels are
pulled before frames are loaded into the channel bank.