mpt: Fix bugs found with the fuzzer (#28)
This fixes two bugs found with the fuzzer. It is now able to run the
fuzzer for a significant period of time without finding bugs. In
addition, this fixes one known bug that I was attempting to find via
fuzzing, but is not able to be found with the test setup.

This adds regression tests for all of the found bugs.